+++ src/Microsoft.Extensions.Caching.Memory/CacheEntry.cs
@@ -30,7 +30,7 @@ namespace Microsoft.Extensions.Caching.Memory
         internal DateTimeOffset? _absoluteExpiration;
         internal TimeSpan? _absoluteExpirationRelativeToNow;
         private TimeSpan? _slidingExpiration;
-        private IDisposable _bookmark;
+        private IDisposable _scopes;
 
         internal readonly object _lock = new object();
 
@@ -58,7 +58,7 @@ namespace Microsoft.Extensions.Caching.Memory
             _notifyCacheEntryDisposed = notifyCacheEntryDisposed;
             _notifyCacheOfExpiration = notifyCacheOfExpiration;
 
-            _bookmark = CacheEntryHelper.EnterScope(this);
+            _scopes = CacheEntryHelper.EnterScope(this);
         }
 
         /// <summary>
@@ -171,7 +171,7 @@ namespace Microsoft.Extensions.Caching.Memory
             if (!_added)
             {
                 _added = true;
-                _bookmark.Dispose();
+                _scopes.Dispose();
                 _notifyCacheEntryDisposed(this);
                 PropageOptions(CacheEntryHelper.Current);
             }
